requirements.txt
langchain>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.1.0
langchain-google-genai>=0.1.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
pdfplumber>=0.10.0
pdf2image>=1.16.3
pytesseract>=0.3.10
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
